+++ src/components/Card/index.tsx
@@ -4,14 +4,14 @@ import Style from './style';
 type CardProps = {
   readonly active: boolean;
   readonly index: number;
-  readonly handleCardClick: (index: number) => void;
+  readonly handleCardMouseDown: (index: number) => void;
 };
 
-const Card = ({ active, index, handleCardClick }: CardProps): JSX.Element =>
+const Card = ({ active, index, handleCardMouseDown }: CardProps): JSX.Element =>
   <Style
     className={`card${active ? ' active' : ''}`}
     onClick={(event: React.MouseEvent<HTMLInputElement>): void => event.stopPropagation()}
-    onMouseDown={(): void => handleCardClick(index)}>
+    onMouseDown={(): void => handleCardMouseDown(index)}>
     Hello
   </Style>;